requirements.txt
apispec==3.1.0
apispec-webframeworks==0.5.2
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorlog==4.0.2
Flask==1.1.1
flask-caching==1.9.0
flask-swagger-ui==3.20.9
Flask-Cors==3.0.8
Flask-Testing==0.7.1
gunicorn==20.0.4
gevent==1.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==2.20.5
marshmallow-mongoengine==0.9.1
mongoengine==0.18.2
psutil==5.6.7
pymongo==3.9.0
pyRFC3339==1.1
pytest==5.3.4
python-dateutil==2.8.0
pytz==2019.3   
PyYAML==5.1.2 
requests==2.21.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
xmltodict==0.12.0
